# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Engineering Supervision System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Projects
...
  /projects/{projectId}/activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List activities for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Create an activity
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityBody"
      responses:
        "201":
          description: Created activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /projects/{projectId}/activities/{id}:
    patch:
      operationId: updateActivity
      tags: [activities]
      summary: Update an activity
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateActivityBody"
      responses:
        "200":
          description: Updated activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete an activity
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Reports
...
  /owner/access/{token}:
    get:
      operationId: getOwnerProjectByToken
      tags: [owner-portal]
      summary: Get project info for owner portal
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project summary for owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerProjectView"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /owner/verify:
    post:
      operationId: verifyOwnerAccess
      tags: [owner-portal]
      summary: Verify owner password for a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerVerifyBody"
      responses:
        "200":
          description: Access granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerProjectView"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Dashboard
...
  /projects/{projectId}/summary:
    get:
      operationId: getProjectSummary
      tags: [dashboard]
      summary: Get project progress summary
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
  /projects/{projectId}/deviation:
    get:
      operationId: getProjectDeviation
      tags: [dashboard]
      summary: Get project deviation analysis
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deviation analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviationAnalysis"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token

    # User
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, engineer, owner]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - fullName
        - email
        - role
        - createdAt
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        ownerEntity:
          type: string
        supervisorEntity:
          type: string
        contractor:
          type: string
        startDate:
          type: string
          format: date
        expectedEndDate:
          type: string
          format: date
        actualEndDate:
          type: ["string", "null"]
          format: date
        status:
          type: string
          enum: [active, completed, delayed, suspended]
        overallProgress:
          type: number
        ownerAccessToken:
          type: ["string", "null"]
        ownerAccessPassword:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - location
        - ownerEntity
        - supervisorEntity
        - contractor
        - startDate
        - expectedEndDate
        - status
        - overallProgress
        - createdAt
        - updatedAt
...
    Activity:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        plannedStartDate:
          type: string
          format: date
        plannedEndDate:
          type: string
          format: date
        actualStartDate:
          type: ["string", "null"]
          format: date
        actualEndDate:
          type: ["string", "null"]
          format: date
        plannedProgress:
          type: number
        actualProgress:
          type: number
        status:
          type: string
          enum: [not_started, in_progress, completed, delayed]
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - name
        - plannedStartDate
        - plannedEndDate
        - plannedProgress
        - actualProgress
        - status
        - sortOrder
        - createdAt
    CreateActivityBody:
      type: object
      properties:
        name:
          type: string
        plannedStartDate:
          type: string
          format: date
        plannedEndDate:
          type: string
          format: date
        actualStartDate:
          type: ["string", "null"]
          format: date
        actualEndDate:
          type: ["string", "null"]
          format: date
        plannedProgress:
          type: number
          default: 0
        actualProgress:
          type: number
          default: 0
        status:
          type: string
          enum: [not_started, in_progress, completed, delayed]
          default: not_started
        sortOrder:
          type: integer
          default: 0
      required:
        - name
        - plannedStartDate
        - plannedEndDate
    UpdateActivityBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        plannedStartDate:
          type: ["string", "null"]
          format: date
        plannedEndDate:
          type: ["string", "null"]
          format: date
        actualStartDate:
          type: ["string", "null"]
          format: date
        actualEndDate:
          type: ["string", "null"]
          format: date
        plannedProgress:
          type: ["number", "null"]
        actualProgress:
          type: ["number", "null"]
        status:
          type: ["string", "null"]
          enum: [not_started, in_progress, completed, delayed, null]
        sortOrder:
          type: ["integer", "null"]

    # Reports
    Report:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        type:
          type: string
          enum: [weekly, monthly]
        reportDate:
          type: string
          format: date
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        workDescription:
          type: string
        progressPercentage:
          type: number
        technicalNotes:
          type: ["string", "null"]
        recommendations:
          type: ["string", "null"]
        imageUrls:
          type: array
          items:
            type: string
        createdById:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - type
        - reportDate
        - periodStart
        - periodEnd
        - workDescription
        - progressPercentage
        - imageUrls
        - createdAt
        - updatedAt
...
    OwnerVerifyBody:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
      required:
        - token
        - password
    OwnerProjectView:
      type: object
      properties:
        project:
          $ref: "#/components/schemas/Project"
        activities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
        reports:
          type: array
          items:
            $ref: "#/components/schemas/Report"
        summary:
          $ref: "#/components/schemas/ProjectSummary"
      required:
        - project
        - activities
        - reports
        - summary
...
    ProjectSummary:
      type: object
      properties:
        projectId:
          type: integer
        overallProgress:
          type: number
        plannedProgress:
          type: number
        activitiesTotal:
          type: integer
        activitiesCompleted:
          type: integer
        activitiesDelayed:
          type: integer
        daysElapsed:
          type: integer
        totalDays:
          type: integer
        daysRemaining:
          type: integer
        delayDays:
          type: integer
        reportsCount:
          type: integer
        filesCount:
          type: integer
      required:
        - projectId
        - overallProgress
        - plannedProgress
        - activitiesTotal
        - activitiesCompleted
        - activitiesDelayed
        - daysElapsed
        - totalDays
        - daysRemaining
        - delayDays
        - reportsCount
        - filesCount
    DeviationAnalysis:
      type: object
      properties:
        projectId:
          type: integer
        timeDeviation:
          type: number
        progressDeviation:
          type: number
        status:
          type: string
          enum: [on_track, slightly_delayed, significantly_delayed, ahead]
        activitiesAnalysis:
          type: array
          items:
            $ref: "#/components/schemas/ActivityDeviation"
      required:
        - projectId
        - timeDeviation
        - progressDeviation
        - status
        - activitiesAnalysis
    ActivityDeviation:
      type: object
      properties:
        activityId:
          type: integer
        activityName:
          type: string
        plannedProgress:
          type: number
        actualProgress:
          type: number
        deviation:
          type: number
        delayDays:
          type: ["integer", "null"]
      required:
        - activityId
        - activityName
        - plannedProgress
        - actualProgress
        - deviation