# lib/api-zod/src/generated/types/createActivityBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Engineering Supervision System
 * OpenAPI spec version: 0.1.0
 */
import type { CreateActivityBodyStatus } from "./createActivityBodyStatus";

export interface CreateActivityBody {
  name: string;
  plannedStartDate: Date;
  plannedEndDate: Date;
  /** @nullable */
  actualStartDate?: Date | null;
  /** @nullable */
  actualEndDate?: Date | null;
  plannedProgress?: number;
  actualProgress?: number;
  status?: CreateActivityBodyStatus;
  sortOrder?: number;
}