# lib/api-zod/src/generated/types/deviationAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Engineering Supervision System
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityDeviation } from "./activityDeviation";
import type { DeviationAnalysisStatus } from "./deviationAnalysisStatus";

export interface DeviationAnalysis {
  projectId: number;
  timeDeviation: number;
  progressDeviation: number;
  status: DeviationAnalysisStatus;
  activitiesAnalysis: ActivityDeviation[];
}